Adopt scripts/cve-reports in the k8s-workflows repository
Currently, in the scan_images.yaml workflow, we're cloning the
https://github.com/canonical/kubeflow-ci.git repository and run
the ./scripts/cve-reports/send-scan.py script from it. We should move
the script to this repository instead, as that script may be removed
in the future.

This commit copies as-is the ./scipts/cve-reports/* files and a few
other auxiliary required to maintain it. The copied tox.ini and
on_pull_request.yaml have been have been simplified to only contain
what is relevant.

The scan_images.yaml action has been updated to use the local script
instead.

Signed-off-by: Claudiu Belu <claudiu.belu@canonical.com>